star_wars_api/requirements.txt
flask
ipdb
request